refactor: rename github module to repo
